+++ packages/functions/ios/RNFBFunctions/RNFBFunctionsModule.h
@@ -16,8 +16,9 @@
  */
 
 #import <Foundation/Foundation.h>
+#import <React/RCTInvalidating.h>
 #import "NativeRNFBTurboFunctions.h"
 
-@interface RNFBFunctionsModule : NSObject <NativeRNFBTurboFunctionsSpec>
+@interface RNFBFunctionsModule : NSObject <NativeRNFBTurboFunctionsSpec, RCTInvalidating>
 
 @end